Reworked ROM database (#53)
## Description
This PR replaces previous `rom_database.c` implementation with a more
verbose version.

## Motivation and Context
Main motivation was to fix reading data from ROMs with non default
endianness and to streamline adding new entries to the database.
Previous implementation was prone to making errors as there was
necessity to keep multiple arrays in sync. In new implementation each
database entry is just one line. As a bonus there's new field for
describing additional game features, like support for rumble or
controller paks.

## How Has This Been Tested?
Change was tested on the SC64 with small number of commercial and
homebrew ROMs.

## Screenshots
No screenshots

## Types of changes
- [x] Improvement (non-breaking change that adds a new feature)
- [x] Bug fix (fixes an issue)
- [ ] Breaking change (breaking change)
- [ ] Config and build (change in the configuration and build system,
has no impact on code or features)

## Checklist:
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

Signed-off-by: Polprzewodnikowy <sc@mateuszfaderewski.pl>