Tentative rewrite of the docs
I have made a first pass of the rewrite of the docs.

My goal was mostly to make things easier to understand for end users, as my knowledge as a developer is quite limited. This means that most of the docs have been rewritten and expanded in certain areas, while the developer guide received only typo and formatting fixes.

Any comments are welcome, as I'm not a developer (one of the main targets of the SC64) nor a native English-speaker.