Docs proofread/rewrites
Proofread and partial rewrite of the docs.
- Unified the title/section names uppercasing.
- Renamed the `..` links to something more human-readable.
- Resorted some (IMO) redundant headers.
- Added some preliminary texts in ROM Configuration regarding CIC types, save types and TV region types. I'm winging it a bit here, so please recheck this one out and comment on what do you want to say here.
- Added comments to point out where screenshots are needed and other issues that might need to be cleared out.
- Typo fixing, both manual and by running Word's autocorrect feature.
- Rewrote the Background Images section.

I feel two things: first, that the MP3 Player page can only be a stub right now (as I can't access it if it exists, so I can't write about it), and the biggest issues would be that there's so much pages right now, the docs might need some merging into categories, summarizing or resorting things. For example, we have in one section the basic control information (File Browser) while on a very different section we can find the advanced control information (Menu Controls).