Update cheats.c
By @Jhynjhiruu

The way the code checks for the libdragon ipl3 is by reading a value that is known to be a jump instruction in the libultra ipl3
but that area of CIC6106 is scrambled
the person who wrote this code accounted for that when inserting the hook, but forgot about it when checking for the libdragon ipl3
so the check reads 0x80A87642 from the ipl3, correctly determines that it's not a jump instruction, and assumes it must be libdragon
the fix is to apply the same scrambling when CIC6106 is detected also when doing the check.

This also fixes DE cheat types.
And CIC_7102 handling.